+++ Source/KISAddonConfig.cs
@@ -1,8 +1,15 @@
-﻿using KSPDev.ConfigUtils;
+﻿// Kerbal Inventory System
+// Mod's author: KospY (http://forum.kerbalspaceprogram.com/index.php?/profile/33868-kospy/)
+// Module authors: KospY, igor.zavoychinskiy@gmail.com
+// License: Restricted
+
+using KSPDev.ConfigUtils;
 using KSPDev.LogUtils;
+using KSPDev.ModelUtils;
 using KSPDev.Types;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 
 namespace KIS {
@@ -16,12 +23,15 @@ sealed class KISAddonConfig : MonoBehaviour {
   [PersistentField("StackableModule/moduleName", isCollection = true)]
   public readonly static List<string> stackableModules = new List<string>();
 
+  [PersistentField("EquipAliases/alias", isCollection = true)]
+  public readonly static List<string> equipAliases = new List<string>();
+
   [PersistentField("Global/breathableAtmoPressure")]
-  public readonly static float breathableAtmoPressure = 0.5f;
+  public static float breathableAtmoPressure = 0.5f;
 
   [PersistentField("EvaInventory")]
   readonly static PersistentConfigNode evaInventory = new PersistentConfigNode();
-  
+
   [PersistentField("EvaPickup")]
   readonly static PersistentConfigNode evaPickup = new PersistentConfigNode();
 
@@ -53,7 +63,9 @@ sealed class KISAddonConfig : MonoBehaviour {
     public override void StartLoad() {
       // Kerbal parts.
       UpdateEvaPrefab(MaleKerbalEva);
+      UpdateEvaPrefab(MaleKerbalEvaVintage);
       UpdateEvaPrefab(FemaleKerbalEva);
+      UpdateEvaPrefab(FemaleKerbalEvaVintage);
 
       // Set inventory module for every pod with crew capacity.
       DebugEx.Info("Loading pod inventories...");
@@ -63,7 +75,8 @@ sealed class KISAddonConfig : MonoBehaviour {
               || avPart.name == MaleKerbalEvaVintage || avPart.name == FemaleKerbalEvaVintage
               || avPart.name == RdKerbalEva
               || !avPart.partPrefab || avPart.partPrefab.CrewCapacity < 1)) {
-          DebugEx.Fine("Found part with CrewCapacity: {0}", avPart.name);
+          DebugEx.Fine("Found part with crew: {0}, CrewCapacity={1}",
+                       avPart.name, avPart.partPrefab.CrewCapacity);
           AddPodInventories(avPart.partPrefab, avPart.partPrefab.CrewCapacity);
         }
       }
@@ -95,28 +108,39 @@ sealed class KISAddonConfig : MonoBehaviour {
 
   public static void AddPodInventories(Part part, int crewCapacity) {
     for (var i = 0; i < crewCapacity; i++) {
+      var moduleInventory =
+          part.AddModule(typeof(ModuleKISInventory).Name) as ModuleKISInventory;
+      KIS_Shared.AwakePartModule(moduleInventory);
+      moduleInventory.invType = ModuleKISInventory.InventoryType.Pod;
+      DebugEx.Fine("{0}: Add pod inventory to match the capacity", part);
+    }
+    var podInventories = part.Modules.OfType<ModuleKISInventory>()
+        .Where(m => m.invType == ModuleKISInventory.InventoryType.Pod)
+        .ToArray();
+    for (var i = 0; i < podInventories.Length; i++) {
       try {
-        var moduleInventory =
-            part.AddModule(typeof(ModuleKISInventory).Name) as ModuleKISInventory;
-        KIS_Shared.AwakePartModule(moduleInventory);
-        var baseFields = new BaseFieldList(moduleInventory);
+        var baseFields = new BaseFieldList(podInventories[i]);
         baseFields.Load(evaInventory);
-        moduleInventory.podSeat = i;
-        moduleInventory.invType = ModuleKISInventory.InventoryType.Pod;
-        DebugEx.Fine("Pod inventory module(s) for seat {0} loaded successfully", i);
+        podInventories[i].podSeat = i;
+        DebugEx.Fine("{0}: Pod inventory for seat {1} loaded successfully", part, i);
       } catch {
-        DebugEx.Error("Pod inventory module(s) for seat {0} can't be loaded!", i);
+        DebugEx.Error("{0}: Pod inventory module for seat {1} can't be loaded!", part, i);
       }
     }
   }
 
   /// <summary>Load config of EVA modules for the requested part name.</summary>
   static void UpdateEvaPrefab(string partName) {
-    var prefab = PartLoader.getPartInfoByName(partName).partPrefab;
-    if (LoadModuleConfig(prefab, typeof(ModuleKISInventory), evaInventory)) {
-      prefab.GetComponent<ModuleKISInventory>().invType = ModuleKISInventory.InventoryType.Eva;
+    var partInfo = PartLoader.getPartInfoByName(partName);
+    if (partInfo != null ){
+      var prefab = partInfo.partPrefab;
+      if (LoadModuleConfig(prefab, typeof(ModuleKISInventory), evaInventory)) {
+        prefab.GetComponent<ModuleKISInventory>().invType = ModuleKISInventory.InventoryType.Eva;
+      }
+      LoadModuleConfig(prefab, typeof(ModuleKISPickup), evaPickup);
+    } else {
+      DebugEx.Info("Skipping EVA model: {0}. Expansion is not installed.", partName);
     }
-    LoadModuleConfig(prefab, typeof(ModuleKISPickup), evaPickup);
   }
 
   /// <summary>Loads config values for the part's module fro the provided config node.</summary>
@@ -136,6 +160,35 @@ sealed class KISAddonConfig : MonoBehaviour {
     baseFields.Load(node);
     DebugEx.Info("Loaded config for {0} on part {1}", moduleType, p);
     return true;
+  }
+
+  /// <summary>Finds a bone to attach the equippable item to.</summary>
+  /// <param name="root">The transform to start searching from.</param>
+  /// <param name="bonePath">The hierarchy search pattern or a KIS alias.</param>
+  /// <returns>The transform or <c>null</c> if nothing found.</returns>
+  /// <seealso cref="KISAddonConfig.equipAliases"/>
+  public static Transform FindEquipBone(Transform root, string bonePath) {
+    Transform res;
+    if (bonePath.StartsWith("alias", StringComparison.Ordinal)) {
+      res = KISAddonConfig.equipAliases
+          .Select(a => a.Split(new[] {','}, 2))
+          .Where(pair => pair.Length == 2 && pair[0] == bonePath)
+          .Select(pair => Hierarchy.FindTransformByPath(root, pair[1]))
+          .FirstOrDefault(t => t != null);
+      DebugEx.Fine("For alias '{0}' found transform: {1}", bonePath, res);
+    } else {
+      res = Hierarchy.FindTransformByPath(root, bonePath);
+      DebugEx.Fine("For bone path '{0}' found transform: {1}", bonePath, res);
+    }
+    if (res == null) {
+      DebugEx.Error("Cannot find object for EVA item: {0}", bonePath);
+      var modelListing = Hierarchy.ListHirerahcy(root);
+      DebugEx.Fine("The following tree was available:");
+      foreach (var modelPath in modelListing) {
+        DebugEx.Fine(modelPath);
+      }
+    }
+    return res;
   }
 }