+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# 1.16 (October 29th, 2018):
+* [Fix #275] No preview meshes shown.
+
+# 1.15 (October 27th, 2018):
+* [Change] KSP 1.5.* compatibility.
+* [Fix #268] Wrong size detection on the MH expansion parts. READ #273!!!
+* [Fix #269] Equip items and remove helmet actions don't work.
+* [Fix #270] Command seats must not be allowed to have inventory.
+* [Fix #271] In some languages the CCK tags are broken.
+
+# 1.14 (July 18th, 2018):
+* [Change] Migrate to KSPDev Utils 0.37.0. A stability fix.
+
+# 1.13 (July 7th, 2018):
+* [Change] Migrate to x64 profile. The 32-bit game mode is no more supported!
+* [Change] Migrate to KSPDev Utils 0.36.0.
+* [Enhancement] Add Italian localization (IT_it).
+
+# 1.12 (May 6th, 2018):
+* [Fix #257] Helmet doesn't remove on the female kerbal models.
+* [Change] Upgrade to KSPDev Utils v0.33.
+
+# 1.11 (March 26th, 2018):
+* [Fix #248] Broken inventory for the new DLC suits.
+* [Fix #252] Allow setting type of the inventory via config.
+* [Fix #253] Kerbal preview is wrong for the expansion pack suites.
+* [Fix #254] Remove helmet doesn't work on the expansion EVA suites.
+
 # 1.10 (March 8th, 2018):
 * KSP 1.4 support.
 * [Change] Update the pods in the loading screens.